Add getFillColor to USMap.
Move fill color calculation out of USMap and into USRiskMap.
